json-replay/requirements.txt
quixstreams==2.7.0
python-dotenv
pandas